refactor(desktop): consolidate verbosity + provider dot constants
Single-home the forked dot palettes: verbosity dot moves to
settings/verbosity.ts (consumed by VerbositySelect, dropdown-utils,
ModelPicker), provider dot becomes a canonical PROVIDER_DOT in
chat-constants (consumed by ProviderSelect, which gains dark-mode).
Removes the dead VERBOSITY_DOT import in ModelPicker.

ACCEPTED visual change: verbosity 'normal' dot amber to blue.
