+++ apps/desktop/src/features/settings/verbosity.ts
@@ -7,6 +7,12 @@ export const VERBOSITY_LABEL: Record<VerbosityLevel, string> = {
   verbose: 'Verbose',
 };
 
+export const VERBOSITY_DOT: Record<VerbosityLevel, string> = {
+  brief: 'bg-success',
+  normal: 'bg-info',
+  verbose: 'bg-danger',
+};
+
 const VERBOSITY_DIRECTIVE: Record<VerbosityLevel, string> = {
   brief:
     'Output verbosity: BRIEF. Output only what is strictly required to answer or act. No preambles, no recaps, no explanations unless asked. Single short sentence per update; one-line end-of-turn.',
